Retrieve full information from iucn_data based on species names

This function takes a vector of species names, searches for them in the iucn_data dataset, and returns full information about the matched species.

Usage

get_iucn_data(species_names)

Arguments

species_names

A character vector of species names to search for in the iucn_data.

Value

A data.table with full information for each species found in iucn_data. If a species is not found, it will return NA for the respective fields.

Examples

x <- c("Cedrela odorata", "Persea americana",
       "Panthera uncia", "Lynx lynx", "Ara militaris", "Zonotrichia capencis")
get_iucn_data(x)
#>         submitted_name  core_id                 scientific_name  kingdom
#> 1        Ara militaris 22685548  Ara militaris (Linnaeus, 1766) ANIMALIA
#> 2      Cedrela odorata    32292              Cedrela odorata L.  PLANTAE
#> 3            Lynx lynx    12519      Lynx lynx (Linnaeus, 1758) ANIMALIA
#> 4       Panthera uncia    22732 Panthera uncia (Schreber, 1775) ANIMALIA
#> 5     Persea americana 96986556          Persea americana Mill.  PLANTAE
#> 6 Zonotrichia capencis     <NA>                            <NA>     <NA>
#>         phylum         class          order      family    genus
#> 1     CHORDATA          AVES PSITTACIFORMES PSITTACIDAE      Ara
#> 2 TRACHEOPHYTA MAGNOLIOPSIDA     SAPINDALES   MELIACEAE  Cedrela
#> 3     CHORDATA      MAMMALIA      CARNIVORA     FELIDAE     Lynx
#> 4     CHORDATA      MAMMALIA      CARNIVORA     FELIDAE Panthera
#> 5 TRACHEOPHYTA MAGNOLIOPSIDA       LAURALES   LAURACEAE   Persea
#> 6         <NA>          <NA>           <NA>        <NA>     <NA>
#>   specific_epithet scientific_name_authorship taxon_rank infraspecific_epithet
#> 1        militaris           (Linnaeus, 1766)    species                  <NA>
#> 2          odorata                         L.    species                  <NA>
#> 3             lynx           (Linnaeus, 1758)    species                  <NA>
#> 4            uncia           (Schreber, 1775)    species                  <NA>
#> 5        americana                      Mill.    species                  <NA>
#> 6             <NA>                       <NA>       <NA>                  <NA>
#>   taxonomic_status accepted_name_usage_id
#> 1         accepted               22685548
#> 2         accepted                  32292
#> 3         accepted                  12519
#> 4         accepted                  22732
#> 5         accepted               96986556
#> 6             <NA>                     NA
#>                                                                                                                                                                                                          bibliographic_citation
#> 1                                                 BirdLife International 2020. Ara militaris (Linnaeus, 1766). The IUCN Red List of Threatened Species 2020: https://doi.org/10.2305/IUCN.UK.2020-3.RLTS.T22685548A179407584.en
#> 2                                                                Mark, J. & Rivers, M.C. 2017. Cedrela odorata L.. The IUCN Red List of Threatened Species 2017: https://doi.org/10.2305/IUCN.UK.2017-3.RLTS.T32292A68080590.en
#> 3                                        Breitenmoser, U., Breitenmoser-Würsten, C., Lanz, T., von Arx, M., Antonevich, A., Bao, W. & Avgan, B. 2015. Lynx lynx (Linnaeus, 1758). The IUCN Red List of Threatened Species 2015:
#> 4          McCarthy, T., Mallon, D., Jackson, R., Zahler, P. & McCarthy, K. 2017. Panthera uncia (Schreber, 1775). The IUCN Red List of Threatened Species 2017: https://doi.org/10.2305/IUCN.UK.2017-2.RLTS.T22732A50664030.en
#> 5 Wegier, A., Lorea Hernández, F., Contreras, A., Tobón, W. & Mastretta-Yanes, A. 2017. Persea americana Mill.. The IUCN Red List of Threatened Species 2017: https://doi.org/10.2305/IUCN.UK.2017-3.RLTS.T96986556A96986588.en
#> 6                                                                                                                                                                                                                          <NA>
#>                                                    references
#> 1 https://apiv3.iucnredlist.org/api/v3/taxonredirect/22685548
#> 2    https://apiv3.iucnredlist.org/api/v3/taxonredirect/32292
#> 3    https://apiv3.iucnredlist.org/api/v3/taxonredirect/12519
#> 4    https://apiv3.iucnredlist.org/api/v3/taxonredirect/22732
#> 5 https://apiv3.iucnredlist.org/api/v3/taxonredirect/96986556
#> 6                                                        <NA>
#>                                                                                                                                                                                                                          source
#> 1                                                 BirdLife International 2020. Ara militaris (Linnaeus, 1766). The IUCN Red List of Threatened Species 2020: https://doi.org/10.2305/IUCN.UK.2020-3.RLTS.T22685548A179407584.en
#> 2                                                                Mark, J. & Rivers, M.C. 2017. Cedrela odorata L.. The IUCN Red List of Threatened Species 2017: https://doi.org/10.2305/IUCN.UK.2017-3.RLTS.T32292A68080590.en
#> 3                                        Breitenmoser, U., Breitenmoser-Würsten, C., Lanz, T., von Arx, M., Antonevich, A., Bao, W. & Avgan, B. 2015. Lynx lynx (Linnaeus, 1758). The IUCN Red List of Threatened Species 2015:
#> 4          McCarthy, T., Mallon, D., Jackson, R., Zahler, P. & McCarthy, K. 2017. Panthera uncia (Schreber, 1775). The IUCN Red List of Threatened Species 2017: https://doi.org/10.2305/IUCN.UK.2017-2.RLTS.T22732A50664030.en
#> 5 Wegier, A., Lorea Hernández, F., Contreras, A., Tobón, W. & Mastretta-Yanes, A. 2017. Persea americana Mill.. The IUCN Red List of Threatened Species 2017: https://doi.org/10.2305/IUCN.UK.2017-3.RLTS.T96986556A96986588.en
#> 6                                                                                                                                                                                                                          <NA>
#>   threat_status occurrence_status infrasp_tag
#> 1    Vulnerable           Present        <NA>
#> 2    Vulnerable           Present        <NA>
#> 3 Least concern           Present        <NA>
#> 4    Vulnerable           Present        <NA>
#> 5 Least concern           Present        <NA>
#> 6          <NA>              <NA>        <NA>